refactor: consolidate redaction helpers
